fix: macos builds, incompatible js signatures, etc
This primarily resolves issues for building on MacOS, there is a couple
of new targets in the `Makefile`: `clean` which is generally expected
and second `dev-macos` which will determine the proper architecture
and paths for openssl through homebrew. The instructions have been
rewritten based off the updates so that it is more clear for building.

There was a case-sensitivity issue, I would recommend developers
operate with `git config --global core.ignorecase false` and even
better if they utilize a case-sensitive file system as the `line`
component and a few other case-sensitive issues were found.

Lastly, I encountered serveral issues with various signatures, the
signatures ultimately caused a mismatch and therefore I had to alter
the behavior of one call and secondly match up the signatures on the
rest which appear to have no influence on actual behavior.